chore: Move bootstrap into ipfs function and added Ipfs::swarm_listen_on
This function would allow us to call Swarm::listen_on directly. This is mostly for relay usage rather than listening on a specific address. This may change in the future.

Bootstrap is moved into ipfs functions to be able to announce ourselves to DHT. It is not clear on why it was not apart of the functions in the first place.