fix/igd: Poll channel receiver if available (#20)
Using block_in_place with block_on from tokio would block the future, other events, as well as the current thread itself. So instead, we should check to determine of the address is private, return a channel with Option<_>, and poll the receiving channel directly for the `Result`. Once the result has been received to remove it from the map and check it prior to inserting it into "mapping_task" for tracking.