chore: Add timeout in transport config for quic `max_idle_timeout`
When 2 nodes are connected via quic and one disconnects from the other suddenly, it takes by default of 30 seconds before the other node is notify about the disconnection. Lowering the timeout will allow the node to be notified quicker while also adding an option to adjust it in the transport configuration.